Move DSL to utils bundle
