+++ bundles/dev.arcovia.mitigation.utils/src/dev/arcovia/mitigation/utils/dsl/selectors/constant/ConstantDataCharacteristicListSelector.java
@@ -1,0 +1,44 @@
+package dev.arcovia.mitigation.utils.dsl.selectors.constant;
+
+import org.dataflowanalysis.analysis.dsl.selectors.DataCharacteristicListSelector;
+import org.dataflowanalysis.analysis.dsl.selectors.DataCharacteristicsSelector;
+
+import dev.arcovia.mitigation.utils.dsl.nodes.BranchNode;
+import dev.arcovia.mitigation.utils.dsl.nodes.ConjunctionNode;
+import dev.arcovia.mitigation.utils.dsl.nodes.DisjunctionNode;
+
+/**
+ * Wraps a {@link DataCharacteristicListSelector} to add constant data characteristic literals to a {@link BranchNode}.
+ * Handles inversion by creating conjunction or disjunction nodes and recursively adding literals for each data
+ * characteristic in the selector.
+ */
+public class ConstantDataCharacteristicListSelector implements ConstantDataSelector {
+    private final DataCharacteristicListSelector selector;
+
+    /**
+     * Constructs a {@link ConstantDataCharacteristicListSelector} wrapping the given
+     * {@link DataCharacteristicListSelector}.
+     * @param selector the {@link DataCharacteristicListSelector} to wrap
+     */
+    public ConstantDataCharacteristicListSelector(DataCharacteristicListSelector selector) {
+        this.selector = selector;
+    }
+
+    /**
+     * Adds literals to the root node based on the data characteristics of this selector. Creates a conjunction or
+     * disjunction node depending on the inversion flag, and adds corresponding constant data characteristic selectors as
+     * predicates.
+     * @param root the {@link BranchNode} to which literals are added
+     */
+    @Override
+    public void addLiterals(BranchNode root) {
+        var incomingDataNode = selector.isInverted() ? new ConjunctionNode() : new DisjunctionNode();
+        root.addPredicate(incomingDataNode);
+
+        var dataSelectors = selector.getDataCharacteristics()
+                .stream()
+                .map(it -> new ConstantDataCharacteristicSelector(new DataCharacteristicsSelector(null, it, selector.isInverted())))
+                .toList();
+        dataSelectors.forEach(it -> it.addLiterals(incomingDataNode));
+    }
+}
