+++ tests/dev.arcovia.mitigation.ilp.tests/META-INF/MANIFEST.MF
@@ -17,6 +17,7 @@ Require-Bundle:
  org.dataflowanalysis.examplemodels,
  tools.mdsd.library.standalone.initialization;bundle-version="0.3.0",
  org.dataflowanalysis.dfd.datadictionary;bundle-version="4.0.0",
- org.dataflowanalysis.dfd.dataflowdiagram;bundle-version="4.0.0"
+ org.dataflowanalysis.dfd.dataflowdiagram;bundle-version="4.0.0",
+ dev.arcovia.mitigation.utils
 Automatic-Module-Name: dev.arcovia.mitigation.ilp.tests
 Bundle-RequiredExecutionEnvironment: JavaSE-17
