+++ tests/dev.arcovia.mitigation.ilp.tests/src/dev/arcovia/mitigation/ilp/tests/BasicTest.java
@@ -25,7 +25,7 @@ import dev.arcovia.mitigation.ilp.OptimizationManager;
 import dev.arcovia.mitigation.sat.IncomingDataLabel;
 import dev.arcovia.mitigation.sat.Label;
 import dev.arcovia.mitigation.sat.NodeLabel;
-import dev.arcovia.mitigation.sat.dsl.CNFTranslation;
+import dev.arcovia.mitigation.utils.dsl.CNFTranslation;
 import dev.arcovia.mitigation.ilp.EvaluationFunction;
 
 public class BasicTest {
